Rename and move json_input to connector.json.input
