+++ tests/unit/connector/test_connector_factory.py
@@ -20,7 +20,7 @@ from logprep.connector.confluent_kafka.input import (
     ConfluentKafkaInput,
 )
 from logprep.connector.dummy.input import DummyInput
-from logprep.input.json_input import JsonInput
+from logprep.connector.json.input import JsonInput
 from logprep.input.jsonl_input import JsonlInput
 from logprep.connector.confluent_kafka.output import (
     ConfluentKafkaOutput,
