+++ tests/unit/connector/test_json_input.py
@@ -5,7 +5,7 @@ from unittest import mock
 
 import pytest
 from logprep.input.input import CriticalInputError
-from logprep.input.json_input import JsonInput
+from logprep.connector.json.input import JsonInput
 
 
 class DummyError(BaseException):
